ui.py split
